chore: Change import paths to reflect renamed modules
- Changed import paths for analysis_types in the analysis subpackages init.
- Changed import paths in network.py for both setup_logging and analysis_types.
- Changed import path in test_network for analysis_types.
- Added per-file-ignores to ruff config for network.py, so we can use 'G' for var names, as it's common practice for networkx.
- Removed unused var in network.py to satisfy ruff's linter.
