+++ insta_message_analyzer/analysis/__init__.py
@@ -1,9 +1,8 @@
-from .analysis_types import ActivityAnalysisResult, TimeSeriesDict, TimeSeriesKey
+from .analysis_types import ActivityAnalysisResult, NetworkAnalysisResult, TimeSeriesDict, TimeSeriesKey
 from .pipeline import AnalysisPipeline
 from .protocol import AnalysisStrategy
 from .strategies.activity import ActivityAnalysis
 from .strategies.network import NetworkAnalysis
-from .types import ActivityAnalysisResult, NetworkAnalysisResult, TimeSeriesDict, TimeSeriesKey
 from .validation import is_activity_analysis_result, is_time_series_dict
 
 __all__ = [
